refactor(rxDucksMiddleware): completely change all the things
BREAKING CHANGE: Completely revamped usage model. Now you're expected to dispatch an action with an `async` property that is a function that takes the stream of all actions and returns an observable. When you  such a function, an Rx Subscription is returned as a nice-to-have, but it is probably better to compose in the termination of async operations with a `takeUntil` operator.