feat(StateObservable): Add getState method to StateObservable
This would be extremely helpful when migrating to the latest versions of redux-observable. We are in a weird spot where our application relies on RxJS 6, but previously only supported redux-observable@0.x. I understand if this is philosophically against what you'd like for the library, but I thought it might be helpful for other just so that they can migrate incrementally. Maybe a 1.0.0-compat version? Please let me know if there's anything else I can do/provide in the meantime.